feat(decorators/assets): reorder import paths
